[refactor] PetAvatar 기능 개선: Gemini 프로바이더로의 전환 및 이미지 업로드를 통한 아바타 생성 기능 추가, 설정 파일 업데이트
